Changes for test cases
