+++ Example/Tests/ConstantsTest.swift
@@ -1,0 +1,28 @@
+//
+//  ConstantsTest.swift
+//  FullAuthIOSClient_Tests
+//
+//  Created by Monica on 13/08/18.
+//  Copyright © 2018 CocoaPods. All rights reserved.
+//
+
+import XCTest
+import FullAuthIOSClient
+
+class ConstantsTest: XCTestCase {
+    
+    override func setUp() {
+        super.setUp()
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+        
+    }
+
+    func testPerformanceExample() {
+        self.measure {
+        }
+    }
+    
+}
